requirements.txt
# AntiSpin requirements (Python >= 3.10, 64-bit is required)
# If you see "no matching distributions for numpy", your Python is likely <3.10 or not 64-bit.
# NumPy 2.2.6 supports Python 3.10-3.13. :contentReference[oaicite:3]{index=3}

# PyTorch CUDA 12.8 wheels
--index-url https://download.pytorch.org/whl/cu128
--extra-index-url https://pypi.org/simple

torch==2.9.0

# HF stack
transformers==4.57.3
sentence-transformers==5.2.0
accelerate==1.12.0

# core numeric stack (NumPy 2 + pandas compatible)
numpy==2.2.6
scipy==1.15.3
pandas==2.2.2
scikit-learn==1.7.2

# text + config + logging
jieba==0.42.1
pyyaml==6.0.1
tqdm==4.67.1
loguru==0.7.3

# UI + report
plotly>=5.18.0
matplotlib>=3.8.0
jinja2>=3.1.0
streamlit==1.52.2

# tests
pytest==8.3.4
